+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 Below should serve as an "upgrade guide" jumping off point as you are migrating projects to newer versions of this package. Ideally, interface changes will be called out specifically and one should not have to go sifting through git diffs to discern what changed.
 
 - `?.?.?` - *WIP*
+   - External changes:
+      - `expect` no longer is assumed to be in your environment through `jest` and is an explicit dependency now.
+   - Internal changes:
+      - linter update to `1.3.2`
+      - tests for `equalsAny`
+      - tests for `containsAll`
 - `1.1.5`
    - Added methods:
       - `testHarness`